pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-mcp"
version = "0.4.0"
description = "SHELET-compliant cognitive prosthetic for AI agents — 25 stratified MCP skills with structural citation discipline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Mordechai Potash"}]
keywords = ["mcp", "brain", "ai", "conversations", "search", "cognitive", "prosthetic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.0",
    "lancedb>=0.20",
    "duckdb>=1.0",
    "pandas>=2.0",
    "pyarrow>=15.0",
    "pyyaml",
    "rich>=13.0",
    "pytz",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
embed = ["fastembed>=0.5"]
embed-torch = ["sentence-transformers>=2.0"]
summarize = ["anthropic>=0.40", "openai>=1.0"]
dashboard = ["fastapi>=0.115", "uvicorn[standard]>=0.30", "jinja2>=3.1"]
dev = ["pytest>=7.0", "build", "twine", "fastapi>=0.115", "uvicorn[standard]>=0.30", "jinja2>=3.1", "httpx>=0.27"]

[project.scripts]
brain-mcp = "brain_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/mordechaipotash/brain-mcp"
Documentation = "https://brainmcp.dev"
Repository = "https://github.com/mordechaipotash/brain-mcp"

[tool.setuptools.packages.find]
include = ["brain_mcp*"]

[tool.setuptools.package-data]
"brain_mcp.dashboard" = ["templates/**/*.html", "static/**/*"]
"brain_mcp" = ["_prompts/*.txt"]
